+++ src/views/Rankings.vue
@@ -4,7 +4,6 @@ import { useRoute } from 'vue-router';
 import { useCompetitionStore } from '@/stores/competitionsStore';
 import { storeToRefs } from 'pinia';
 import type { Archer, ArcherScore } from '@/types';
-import { generateRankingByClubPDF } from '@/utils/rankingByClubPDF';
 import { generateRankingPDF } from '@/utils/rankingPDF';
 import type { RankingCategory } from '@/types/ranking';
 import { 
@@ -215,38 +214,6 @@ async function generatePDF() {
   }
 }
 
-async function generateClubPDF() {
-  if (!competition.value || isGeneratingPDF.value) return; 
-  
-  try {
-      isGeneratingPDF.value = true;
-      showExportModal.value = false;
-    
-    // Generate the PDF
-  const pdfBytes = await generateRankingByClubPDF(
-    competition.value,
-    archers.value,
-  );
-
-  const blob = new Blob([pdfBytes], { type: 'application/pdf' });
-  const url = URL.createObjectURL(blob);
-  const link = document.createElement('a');
-  link.href = url;
-  link.download = `Placement des archers par club.pdf`;
-  document.body.appendChild(link);
-  link.click();
-    
-
-  URL.revokeObjectURL(url);
-  document.body.removeChild(link);
-  } catch (error) {
-    console.error('Erreur lors de la génération du PDF:', error);
-    alert('Une erreur est survenue lors de la génération du PDF. Veuillez réessayer.');
-    } finally {
-      isGeneratingPDF.value = false;
-    }
-  }
-
 </script>
 
 <template>
@@ -263,14 +230,6 @@ async function generateClubPDF() {
           >
             <DocumentArrowDownIcon class="w-5 h-5" />
             {{ isGeneratingPDF ? 'Génération du PDF...' : 'Exporter en PDF' }}
-          </button>
-
-          <button 
-            class="btn btn-primary flex items-center gap-2"
-            @click="generateClubPDF"
-          >
-            <DocumentArrowDownIcon class="w-5 h-5" />
-            {{ 'Placement par club' }}
           </button>
         </div>
